refactor: combine peer page & calendar veiw page, replace enter peer name to random peer name
