--- app/page.module.css
@@ -1,13 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  margin-top: 2rem;
-  max-width: 22rem;
-}
-
-.inputForm {
-  display: flex;
-  flex-direction: row;
-  margin-top: 0.5rem;
-  box-shadow: 0 0 1rem 0.5rem rgb(0 0 0 / 10%);
-}
