+++ app/styles/page.module.css
@@ -1,3 +1,7 @@
+.main {
+  width: 340px;
+}
+
 .textArea {
   width: 100%;
   height: 8rem;
